+++ client/src/features/editor/Editor.tsx
@@ -7,8 +7,6 @@ import { BlockId } from "@noctaCrdt/NodeId";
 import {
   RemoteCharInsertOperation,
   serializedEditorDataProps,
-  TextColorType,
-  BackgroundColorType,
 } from "node_modules/@noctaCrdt/Interfaces.ts";
 import { useRef, useState, useCallback, useEffect, useMemo } from "react";
 import { useSocketStore } from "@src/stores/useSocketStore.ts";
@@ -22,6 +20,7 @@ import {
 import { Block } from "./components/block/Block";
 import { useBlockDragAndDrop } from "./hooks/useBlockDragAndDrop";
 import { useBlockOptionSelect } from "./hooks/useBlockOption";
+import { useCopyAndPaste } from "./hooks/useCopyAndPaste.ts";
 import { useMarkdownGrammer } from "./hooks/useMarkdownGrammer";
 import { useTextOptionSelect } from "./hooks/useTextOptions.ts";
 import { getTextOffset } from "./utils/domSyncUtils.ts";
@@ -37,12 +36,7 @@ interface EditorProps {
   serializedEditorData: serializedEditorDataProps;
   pageTitle: string;
 }
-interface ClipboardMetadata {
-  value: string;
-  style: string[];
-  color: TextColorType | undefined;
-  backgroundColor: BackgroundColorType | undefined;
-}
+
 export const Editor = ({ onTitleChange, pageId, pageTitle, serializedEditorData }: EditorProps) => {
   const {
     sendCharInsertOperation,
@@ -115,6 +109,12 @@ export const Editor = ({ onTitleChange, pageId, pageTitle, serializedEditorData 
       pageId,
     },
   );
+
+  const { handleCopy, handlePaste } = useCopyAndPaste({
+    editorCRDT: editorCRDT.current,
+    setEditorState,
+    pageId,
+  });
 
   const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const newTitle = e.target.value;
@@ -267,105 +267,6 @@ export const Editor = ({ onTitleChange, pageId, pageTitle, serializedEditorData 
     } else {
       onKeyDown(e);
     }
-  };
-
-  const handleCopy = (
-    e: React.ClipboardEvent<HTMLDivElement>,
-    blockRef: HTMLDivElement | null,
-    block: CRDTBlock,
-  ) => {
-    e.preventDefault();
-    if (!blockRef) return;
-
-    const selection = window.getSelection();
-    if (!selection || selection.isCollapsed || !blockRef) return;
-
-    const range = selection.getRangeAt(0);
-    if (!blockRef.contains(range.commonAncestorContainer)) return;
-
-    // 선택된 텍스트의 시작과 끝 위치 계산
-    const startOffset = getTextOffset(blockRef, range.startContainer, range.startOffset);
-    const endOffset = getTextOffset(blockRef, range.endContainer, range.endOffset);
-
-    // 선택된 텍스트와 스타일 정보 추출
-    const selectedChars = block.crdt.LinkedList.spread().slice(startOffset, endOffset);
-
-    // 커스텀 데이터 포맷으로 저장
-    const customData = {
-      text: selectedChars.map((char) => char.value).join(""),
-      metadata: selectedChars.map(
-        (char) =>
-          ({
-            value: char.value,
-            style: char.style,
-            color: char.color,
-            backgroundColor: char.backgroundColor,
-          }) as ClipboardMetadata,
-      ),
-    };
-
-    // 일반 텍스트와 커스텀 데이터 모두 클립보드에 저장
-    e.clipboardData.setData("text/plain", customData.text);
-    e.clipboardData.setData("application/x-nocta-formatted", JSON.stringify(customData));
-  };
-
-  const handlePaste = (e: React.ClipboardEvent<HTMLDivElement>, block: CRDTBlock) => {
-    e.preventDefault();
-
-    const customData = e.clipboardData.getData("application/x-nocta-formatted");
-
-    if (customData) {
-      const { metadata } = JSON.parse(customData);
-      const caretPosition = block.crdt.currentCaret;
-
-      metadata.forEach((char: ClipboardMetadata, index: number) => {
-        const insertPosition = caretPosition + index;
-        const charNode = block.crdt.localInsert(
-          insertPosition,
-          char.value,
-          block.id,
-          pageId,
-          char.style,
-          char.color,
-          char.backgroundColor,
-        );
-        sendCharInsertOperation({
-          node: charNode.node,
-          blockId: block.id,
-          pageId,
-          style: char.style,
-          color: char.color,
-          backgroundColor: char.backgroundColor,
-        });
-      });
-
-      editorCRDT.current.currentBlock!.crdt.currentCaret = caretPosition + metadata.length;
-    } else {
-      const text = e.clipboardData.getData("text/plain");
-
-      if (!block || text.length === 0) return;
-
-      const caretPosition = block.crdt.currentCaret;
-
-      // 텍스트를 한 글자씩 순차적으로 삽입
-      text.split("").forEach((char, index) => {
-        const insertPosition = caretPosition + index;
-        const charNode = block.crdt.localInsert(insertPosition, char, block.id, pageId);
-        sendCharInsertOperation({
-          node: charNode.node,
-          blockId: block.id,
-          pageId,
-        });
-      });
-
-      // 캐럿 위치 업데이트
-      editorCRDT.current.currentBlock!.crdt.currentCaret = caretPosition + text.length;
-    }
-
-    setEditorState({
-      clock: editorCRDT.current.clock,
-      linkedList: editorCRDT.current.LinkedList,
-    });
   };
 
   const handleCompositionEnd = useCallback(
